Moving classes and Javadoc 	- moved BotController to intelligence 	- moved MoveNode, Setting, TreeHeader to new directory "datastructures" in "intelligence" 	- added even more Javadoc 	- added double-whitespace formatting again
